requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
a2wsgi==1.10.7
    # via bentoml
accelerate==1.2.1
    # via transformers
aiohappyeyeballs==2.4.4
    # via aiohttp
aiohttp==3.11.11
    # via bentoml
aiosignal==1.3.2
    # via aiohttp
aiosqlite==0.20.0
    # via bentoml
annotated-types==0.7.0
    # via pydantic
anyio==4.7.0
    # via
    #   httpx
    #   httpx-ws
    #   starlette
    #   watchfiles
appdirs==1.4.4
    # via fs
asgiref==3.8.1
    # via opentelemetry-instrumentation-asgi
attrs==24.3.0
    # via
    #   aiohttp
    #   bentoml
    #   cattrs
bentoml==1.3.16
    # via image-gen-service (pyproject.toml)
cattrs==23.1.2
    # via bentoml
certifi==2024.12.14
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.0
    # via requests
circus==0.18.0
    # via bentoml
click==8.1.8
    # via
    #   bentoml
    #   click-option-group
    #   uvicorn
click-option-group==0.5.6
    # via bentoml
cloudpickle==3.1.0
    # via bentoml
deepmerge==2.0
    # via bentoml
deprecated==1.2.15
    # via opentelemetry-api
diffusers==0.31.0
    # via image-gen-service (pyproject.toml)
filelock==3.16.1
    # via
    #   diffusers
    #   huggingface-hub
    #   torch
    #   transformers
    #   triton
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
fs==2.4.16
    # via bentoml
fsspec==2024.12.0
    # via
    #   huggingface-hub
    #   torch
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
    #   wsproto
httpcore==1.0.7
    # via
    #   httpx
    #   httpx-ws
httpx==0.28.1
    # via
    #   bentoml
    #   httpx-ws
httpx-ws==0.7.0
    # via bentoml
huggingface-hub==0.27.0
    # via
    #   accelerate
    #   diffusers
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==6.11.0
    # via
    #   diffusers
    #   opentelemetry-api
inflection==0.5.1
    # via bentoml
jinja2==3.1.5
    # via
    #   bentoml
    #   torch
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
networkx==3.2.1
    # via torch
numpy==2.0.2
    # via
    #   accelerate
    #   bentoml
    #   diffusers
    #   transformers
nvidia-cublas-cu12==12.4.5.8
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.4.127
    # via torch
nvidia-cuda-nvrtc-cu12==12.4.127
    # via torch
nvidia-cuda-runtime-cu12==12.4.127
    # via torch
nvidia-cudnn-cu12==9.1.0.70
    # via torch
nvidia-cufft-cu12==11.2.1.3
    # via torch
nvidia-curand-cu12==10.3.5.147
    # via torch
nvidia-cusolver-cu12==11.6.1.9
    # via torch
nvidia-cusparse-cu12==12.3.1.170
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-ml-py==12.560.30
    # via bentoml
nvidia-nccl-cu12==2.21.5
    # via torch
nvidia-nvjitlink-cu12==12.4.127
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.4.127
    # via torch
opentelemetry-api==1.20.0
    # via
    #   bentoml
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-aiohttp-client
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-sdk
opentelemetry-instrumentation==0.41b0
    # via
    #   bentoml
    #   opentelemetry-instrumentation-aiohttp-client
    #   opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-aiohttp-client==0.41b0
    # via bentoml
opentelemetry-instrumentation-asgi==0.41b0
    # via bentoml
opentelemetry-sdk==1.20.0
    # via bentoml
opentelemetry-semantic-conventions==0.41b0
    # via
    #   bentoml
    #   opentelemetry-instrumentation-aiohttp-client
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-sdk
opentelemetry-util-http==0.41b0
    # via
    #   bentoml
    #   opentelemetry-instrumentation-aiohttp-client
    #   opentelemetry-instrumentation-asgi
packaging==24.2
    # via
    #   accelerate
    #   bentoml
    #   huggingface-hub
    #   pip-requirements-parser
    #   transformers
pathspec==0.12.1
    # via bentoml
pillow==11.0.0
    # via diffusers
pip-requirements-parser==32.0.1
    # via bentoml
prometheus-client==0.21.1
    # via bentoml
prompt-toolkit==3.0.36
    # via questionary
propcache==0.2.1
    # via
    #   aiohttp
    #   yarl
protobuf==5.29.2
    # via image-gen-service (pyproject.toml)
psutil==6.1.1
    # via
    #   accelerate
    #   bentoml
    #   circus
pydantic==2.10.4
    # via bentoml
pydantic-core==2.27.2
    # via pydantic
pygments==2.18.0
    # via rich
pyparsing==3.2.0
    # via pip-requirements-parser
python-dateutil==2.9.0.post0
    # via bentoml
python-dotenv==1.0.1
    # via bentoml
python-json-logger==3.2.1
    # via bentoml
python-multipart==0.0.20
    # via bentoml
pyyaml==6.0.2
    # via
    #   accelerate
    #   bentoml
    #   huggingface-hub
    #   transformers
pyzmq==26.2.0
    # via circus
questionary==2.0.1
    # via bentoml
regex==2024.11.6
    # via
    #   diffusers
    #   transformers
requests==2.32.3
    # via
    #   diffusers
    #   huggingface-hub
    #   transformers
rich==13.9.4
    # via bentoml
safetensors==0.4.5
    # via
    #   accelerate
    #   diffusers
    #   transformers
schema==0.7.7
    # via bentoml
sentencepiece==0.2.0
    # via image-gen-service (pyproject.toml)
setuptools==75.6.0
    # via
    #   fs
    #   opentelemetry-instrumentation
    #   torch
simple-di==0.1.5
    # via bentoml
six==1.17.0
    # via
    #   fs
    #   python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.42.0
    # via bentoml
sympy==1.13.1
    # via torch
tokenizers==0.21.0
    # via transformers
tomli-w==1.1.0
    # via bentoml
torch==2.5.1
    # via
    #   image-gen-service (pyproject.toml)
    #   accelerate
    #   transformers
torchao==0.7.0
    # via image-gen-service (pyproject.toml)
tornado==6.4.2
    # via circus
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   transformers
transformers==4.47.1
    # via image-gen-service (pyproject.toml)
triton==3.1.0
    # via torch
typing-extensions==4.12.2
    # via
    #   aiosqlite
    #   anyio
    #   huggingface-hub
    #   opentelemetry-sdk
    #   pydantic
    #   pydantic-core
    #   torch
urllib3==2.3.0
    # via requests
uv==0.5.11
    # via bentoml
uvicorn==0.34.0
    # via bentoml
watchfiles==1.0.3
    # via bentoml
wcwidth==0.2.13
    # via prompt-toolkit
wrapt==1.17.0
    # via
    #   deprecated
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-aiohttp-client
wsproto==1.2.0
    # via httpx-ws
yarl==1.18.3
    # via aiohttp
zipp==3.21.0
    # via importlib-metadata
